pyproject.toml
[project]
name = "jira-scripts"
version = "0.1.0"
description = "Simple JIRA data collection scripts"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
